+++ src/main/java/com/sivalabs/modernboot/models/User.java
@@ -1,8 +1,9 @@
 package com.sivalabs.modernboot.models;
 
-public record User(Integer id,
-                   String name,
-                   String username,
-                   String email,
-                   String phone,
-                   String website) {}
+public record User(
+       Long id,
+       String name,
+       String username,
+       String email,
+       String phone,
+       String website) {}
